test(browser): combine host retention and guest recovery
